feat(react): export most utils (#3297)
