+++ .changeset/bright-deers-end.md
@@ -1,0 +1,5 @@
+---
+"@floating-ui/react": patch
+---
+
+feat: export most utils under `@floating-ui/react/utils`. This makes it a lot easier to write your own custom interaction hooks, or copy/paste the existing ones to change the way they are implemented yourself without patching or waiting for features/breaking changes.
