+++ packages/react/src/_deprecated-inner.ts
@@ -1,10 +1,9 @@
-import {getUserAgent} from '@floating-ui/react/utils';
-import {evaluate, max, min, round} from '@floating-ui/utils';
-import {detectOverflow, offset, type Derivable} from '@floating-ui/react-dom';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
+import {detectOverflow, offset, type Derivable} from '@floating-ui/react-dom';
+import {evaluate, max, min, round} from '@floating-ui/utils';
+import {useEffectEvent, getUserAgent} from '@floating-ui/react/utils';
 
-import {useEffectEvent} from './hooks/utils/useEffectEvent';
 import type {
   DetectOverflowOptions,
   ElementProps,
