+++ packages/react/src/hooks/useClick.ts
@@ -1,9 +1,9 @@
+import * as React from 'react';
+import {isHTMLElement} from '@floating-ui/utils/dom';
 import {
   isMouseLikePointerType,
   isTypeableElement,
 } from '@floating-ui/react/utils';
-import {isHTMLElement} from '@floating-ui/utils/dom';
-import * as React from 'react';
 
 import type {ElementProps, FloatingRootContext} from '../types';
