+++ packages/react/src/hooks/useClientPoint.ts
@@ -1,14 +1,14 @@
+import * as React from 'react';
+import {getWindow} from '@floating-ui/utils/dom';
 import {
   contains,
   getTarget,
   isMouseLikePointerType,
+  useModernLayoutEffect,
+  useEffectEvent,
 } from '@floating-ui/react/utils';
-import {getWindow} from '@floating-ui/utils/dom';
-import * as React from 'react';
-import useModernLayoutEffect from 'use-isomorphic-layout-effect';
 
 import type {ContextData, ElementProps, FloatingRootContext} from '../types';
-import {useEffectEvent} from './utils/useEffectEvent';
 
 function createVirtualElement(
   domElement: Element | null | undefined,
