+++ packages/react/src/hooks/useFloating.ts
@@ -1,10 +1,10 @@
+import * as React from 'react';
 import {
   useFloating as usePosition,
   type VirtualElement,
 } from '@floating-ui/react-dom';
 import {isElement} from '@floating-ui/utils/dom';
-import * as React from 'react';
-import useModernLayoutEffect from 'use-isomorphic-layout-effect';
+import {useModernLayoutEffect} from '@floating-ui/react/utils';
 
 import {useFloatingTree} from '../components/FloatingTree';
 import type {
