--- packages/react/src/hooks/utils/useLatestRef.ts
@@ -1,10 +1,0 @@
-import {useRef} from 'react';
-import useModernLayoutEffect from 'use-isomorphic-layout-effect';
-
-export function useLatestRef<T>(value: T) {
-  const ref = useRef<T>(value);
-  useModernLayoutEffect(() => {
-    ref.current = value;
-  });
-  return ref;
-}
