+++ packages/react/src/utils/hooks.ts
@@ -1,5 +1,16 @@
 import * as React from 'react';
-import {SafeReact} from '../../utils/safeReact';
+import useModernLayoutEffect from 'use-isomorphic-layout-effect';
+import {SafeReact} from './safeReact';
+
+export {useModernLayoutEffect};
+
+export function useLatestRef<T>(value: T) {
+  const ref = React.useRef<T>(value);
+  useModernLayoutEffect(() => {
+    ref.current = value;
+  });
+  return ref;
+}
 
 type AnyFunction = (...args: any[]) => any;
