+++ packages/react/src/utils/markOthers.ts
@@ -1,7 +1,7 @@
 // Modified to add conditional `aria-hidden` support:
 // https://github.com/theKashey/aria-hidden/blob/9220c8f4a4fd35f63bee5510a9f41a37264382d4/src/index.ts
-import {getDocument} from '@floating-ui/react/utils';
 import {getNodeName} from '@floating-ui/utils/dom';
+import {getDocument} from './element';
 
 type Undo = () => void;
